prefer go1.16's embedded templates
Go has first class support for file embedding since 1.16,
so we can take advantage of it to factor out inline content.

As it has nice integration with html/template, we also
replace our ad-hoc string streaming with actual templates.

I feel like the code is somewhat cleaner as a result, but
this was mostly an excuse to try out the new embed and template
stuff in go1.16.