allow resizing images to enclosing slide
I'd like for images to automatically scale down
to the size of their enclosing slide, particularly
when multiple images appear side-by-side.

However, as `:has(> img)` is not yet implemented
in most browsers, writing a stylesheet for this
seems almost impossible as long as the slide
is structured as `section>p>img`.

I could write a proper blackfriday parser for this,
but a few horrible regexes will probably do for now.
The resulting `section>img` is far easier to style.

As a bonus, we can now also annotate slides
with arbitrary css classes, which seems useful.