allow hot-reloading presentation without js
Thanks to the innovations of the Netscape era, we can
reload our presentation without any js whatsoever.

The ideal solution would predeclare a HTTP `Trailers: Refresh`
and then conditionally set a `Refresh` trailer if we hit the
refresh codepath. However, although in the standards,
no modern browser seems to honor trailers these days.

Our workaround is to send the Refresh header on every request,
unless we're shutting down. This means that we can simply
close the connection to force the client to re-connect
and grab the fresh version of the slides in the refresh case.

However, this also means that when shutting down we'll need
to anticipate a wave of final refreshes, and continue to
serve files for a bit longer after the `quit` was issued.
This shouldn't produce too much load, as the big assets
are most likely cached at the client already.